Grid render "optimization"
That's the last TODO down, but this was my mistake. I should benchmark before optimizing stuff because this "optimized" version is actually slower than the previous code. Turns out the grid render code runs really fast already, any "optimizations" is likely to make it worse. Just wanted to get this into git's history because it took some hours to get it right and might be useful elsewhere