Custom external map functionality for local network multiplayer mode
Used a Ninepatch for the playmode stage, so I think it would be faster? At least it should look better now. Also, changed version to v0.0.3-dev to show that this is a development version and avoid clashes with the beta versions. Tested the networking, but might have bugs, hard to test every possible scenario. A bit of cleanup on the code coming soon